requirements.txt
scikit-learn
numpy
pandas==2.0.0
matplotlib
typing
seaborn
pymongo[srv] == 4.2.0
pyyaml
scipy==1.10.1
#-e .